recipient account view inc. QRCode, balance and profile
